Fix dependency loops
