+++ packages/web3-providers-base/package.json
@@ -23,6 +23,7 @@
         "@types/jest": "^26.0.22",
         "@typescript-eslint/eslint-plugin": "^4.22.0",
         "@typescript-eslint/parser": "^4.22.0",
+        "axios": "^0.21.1",
         "eslint": "^7.24.0",
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-import": "^2.22.1",
@@ -30,8 +31,6 @@
         "prettier": "^2.3.0",
         "ts-jest": "^26.5.4",
         "typescript": "^4.2.4",
-        "web3-providers-http": "^4.0.0",
-        "web3-eth": "^4.0.0",
         "web3-utils": "^4.0.0"
     }
 }
