+++ packages/web3-providers-base/types.ts
@@ -1,7 +1,15 @@
 import { EventEmitter } from 'events';
-import { HttpOptions } from 'web3-providers-http/types';
-import { EthFilter, EthTransaction } from 'web3-eth/types';
-import { PrefixedHexString, ValidTypesEnum } from 'web3-utils/types';
+import {
+    PrefixedHexString,
+    ValidTypes,
+    ValidTypesEnum,
+    EthFilter,
+    EthTransaction,
+    EthMinedTransaction,
+    EthLog,
+    CompiledSolidity
+} from 'web3-utils/types';
+import { AxiosRequestConfig } from 'axios';
 
 export type ProviderCallOptions = HttpOptions | undefined; // HttpOptions | WsOptions | IpcOptions
 export type RpcParams = (
@@ -14,6 +22,15 @@ export type RpcParams = (
 
 export interface ProviderOptions {
     providerUrl: string;
+}
+
+export interface HttpOptions {
+    axiosConfig?: AxiosRequestConfig;
+    subscriptionOptions?: SubscriptionOptions;
+}
+
+export interface SubscriptionOptions {
+    milisecondsBetweenRequests?: number;
 }
 
 export interface RpcOptions {
@@ -43,6 +60,109 @@ export interface RpcResponse {
     id: number;
     jsonrpc: string;
     result: any;
+}
+
+export interface RpcStringResult extends RpcResponse {
+    result: string;
+}
+
+export interface RpcPrefixedHexStringResult extends RpcResponse {
+    result: PrefixedHexString;
+}
+
+export interface RpcValidTypeResult extends RpcResponse {
+    result: ValidTypes;
+}
+
+export interface RpcBooleanResult extends RpcResponse {
+    result: boolean;
+}
+
+export interface RpcSyncingResult extends RpcResponse {
+    result:
+        | boolean
+        | {
+              startingBlock: ValidTypes;
+              currentBlock: ValidTypes;
+              highestBlock: ValidTypes;
+          };
+}
+
+export interface RpcAccountsResult extends RpcResponse {
+    result: PrefixedHexString[];
+}
+
+export interface RpcBlockResult extends RpcResponse {
+    result: null | {
+        number: ValidTypes | null;
+        hash: PrefixedHexString | null;
+        parentHash: PrefixedHexString;
+        nonce: PrefixedHexString | null;
+        sha3Uncles: PrefixedHexString;
+        logsBloom: PrefixedHexString | null;
+        transactionsRoot: PrefixedHexString;
+        stateRoot: PrefixedHexString;
+        receiptsRoot: PrefixedHexString;
+        miner: PrefixedHexString;
+        difficulty: ValidTypes;
+        totalDifficulty: ValidTypes;
+        extraData: PrefixedHexString;
+        size: ValidTypes;
+        gasLimit: ValidTypes;
+        gasUsed: ValidTypes;
+        timestamp: ValidTypes;
+        transactions: EthMinedTransaction[] | PrefixedHexString[];
+        uncles: PrefixedHexString[];
+    };
+}
+
+export interface RpcTransactionResult extends RpcResponse {
+    result: null | {
+        blockHash: PrefixedHexString | null;
+        blockNumber: ValidTypes | null;
+        from: PrefixedHexString;
+        gas: ValidTypes;
+        gasPrice: ValidTypes;
+        hash: PrefixedHexString;
+        input: PrefixedHexString;
+        nonce: ValidTypes;
+        to: PrefixedHexString | null;
+        transactionIndex: ValidTypes | null;
+        value: ValidTypes;
+        v: ValidTypes;
+        r: PrefixedHexString;
+        s: PrefixedHexString;
+    };
+}
+
+export interface RpcTransactionReceiptResult extends RpcResponse {
+    result: null | {
+        transactionHash: PrefixedHexString;
+        transactionIndex: ValidTypes;
+        blockHash: PrefixedHexString;
+        blockNumber: ValidTypes;
+        from: PrefixedHexString;
+        to: PrefixedHexString | null;
+        cumulativeGasUsed: ValidTypes;
+        gasUsed: ValidTypes;
+        contractAddress: PrefixedHexString | null;
+        logs: EthLog[];
+        logsBloom: PrefixedHexString;
+        root?: PrefixedHexString;
+        status?: ValidTypes;
+    };
+}
+
+export interface RpcStringArrayResult extends RpcResponse {
+    result: string[];
+}
+
+export interface RpcCompiledSolidityResult extends RpcResponse {
+    result: CompiledSolidity;
+}
+
+export interface RpcLogResult extends RpcResponse {
+    result: EthLog[];
 }
 
 export interface SubscriptionResponse {
