--- packages/web3-providers-http/types.ts
@@ -1,10 +1,0 @@
-import { AxiosRequestConfig } from 'axios';
-
-export interface HttpOptions {
-    axiosConfig?: AxiosRequestConfig;
-    subscriptionOptions?: SubscriptionOptions;
-}
-
-export interface SubscriptionOptions {
-    milisecondsBetweenRequests?: number;
-}
