+++ packages/web3-utils/package.json
@@ -29,8 +29,7 @@
         "jest": "^26.6.3",
         "prettier": "^2.3.0",
         "ts-jest": "^26.5.4",
-        "typescript": "^4.2.4",
-        "web3-providers-base": "^1.0.0"
+        "typescript": "^4.2.4"
     },
     "dependencies": {
         "ethereumjs-util": "^7.0.10"
